+++ docs/02_guides/code/scrapy_project/src/settings.py
@@ -1,0 +1,8 @@
+BOT_NAME = 'titlebot'
+DEPTH_LIMIT = 1
+LOG_LEVEL = 'INFO'
+NEWSPIDER_MODULE = 'src.spiders'
+ROBOTSTXT_OBEY = True
+SPIDER_MODULES = ['src.spiders']
+TELNETCONSOLE_ENABLED = False
+TWISTED_REACTOR = 'twisted.internet.asyncioreactor.AsyncioSelectorReactor'